Refactor ParametersMixin into OptionsManager
Will now be used more generally than just in solver options setting.
